Have the Blaze client robustly determine if the server OOM'd.
The new approach is that the Blaze client tells the server's JVM to create a special file before it throws an OOM; then if the Blaze client is unable to otherwise determine the exit code to use (i.e. gRPC connection broken and the server was unable to write an abrupt exit file), it uses the presence of this file to determine it should use `33`.

By having the mechanism be solely in client-land and at the lowest possible level in the JVM, we're no longer at the mercy of the server's app land code working correctly and arbitrary code outside our control not killing JVM before our existing mitigations (like the abrupt exit file) kick in.

The new approach as-is improves quality but doesn't improve speed. As an optional future improvement, we can have the invoker of Blaze notice the write of the special file as soon as it happens and immediately use an exit code of `33`.

PiperOrigin-RevId: 722716155
Change-Id: I00d0f9435bc812678703cd7caf29e20b6351f185