Configure `--run_under` target for the test exec platform
With `--incompatible_bazel_test_exec_run_under`, the `--run_under` target should be configured for the execution platform of the test action, not the default execution platform of the test rule.

This is still not fully correct if `testing.ExecutionInfo` is used to specify an alternative test exec group, but that's a far more difficult change: it would potentially involve a split transition with one configuration for each exec group.

Work towards #23179

Closes #25177.

PiperOrigin-RevId: 722802891
Change-Id: Ie168bf496d46645f7bd2ad739859504c65286d20