[8.1.0] Bazel's test XML parser now supports scientific e notation for time durations. (#25186)
Technically strings like "1e2" are not valid for the `time` attribute
(that
attribute's schema is supposed to be `xs:decimal`), however some JUNIT
writers
are non-compliant and use scientific e notation and we want Bazel to
still work
with them.

While I'm here, improve the code comments (the existing comment about us
having
to support values without decimal points like "12" as milliseconds is
backwards as
noted in
https://github.com/bazelbuild/bazel/issues/24605#issuecomment-2524851929)
and add tests.

Fixes #24605.

PiperOrigin-RevId: 722963681
Change-Id: I9383cf7435ae7843f07a5abddbcf790f275403d0

Commit
https://github.com/bazelbuild/bazel/commit/f52568f1d5fa15538d2f7c99401417d2cd24679c

Co-authored-by: Googler <nharmata@google.com>