Evaluate top-level aspects' required providers before creating the aspect key
if the rule does not satisfy the aspect required provider, its evaluation is dropped in the AspectFunction. This change moves this check before creating the aspect key in preparation for [Aspect Propagation Controller](https://docs.google.com/document/d/1LfglQ4naytDXmvT2Y8ptnoWOv8awhuaToNGmD2dRGdU/edit?tab=t.0) that will enable running a starlark function to determine if the AspectKey should be created.

This change also renamed `BuildTopLevelAspectsDetailsFunction` to `LoadTopLevelAspectsFunction` and made it focus only on loading the top-level aspects and returning a list of `Aspect` so that it can be done once for all top-level targets.

PiperOrigin-RevId: 725261454
Change-Id: If6732a13174357e60c71d2f6ca1de272e67f75c8