Introduce native.package_default_visibility, which returns the default visibility for the package.
This is useful for creating symbolic macros that depend on targets that are created by another macro and are not passed in as an attribute (thus relying on visibility). This allows the creating macro to add visibility to the consuming macro's definition location, without overriding the default visibility for the package where the targets are created.

PiperOrigin-RevId: 725364901
Change-Id: I57ff3fcfa9792745efa42ac756f45679ebc03c60