Don't fetch empty digest on failure status in `FetchBlobResponse`
When the `GrpcRemoteDownloader` receives a `FetchBlobResponse` with an error status (instead of a connection level status error), it must fail and fall back to the HTTP downloader if configured instead of ignoring the status and attempting to fetch the (usually empty) digest from the cache. The previous behavior could result in errors being masked and resulting in an empty file being downloaded if the remote cache accepts empty `Digest` messages.