Remove redundant logic from PersistentMap.load().
Calling loadEntries(false) already suppresses exceptions; we don't need to suppress them a second time.

The updated version can now throw if the file is concurrently mutated (e.g. in between the size check and the read in loadEntries), but that's not something we need to worry about (and I don't see why we'd want to worry about it for the journal, but not the main file).

PiperOrigin-RevId: 725630121
Change-Id: Ic4cbec5e289e1cf0ec0ed95ac8f913d7bc4b74c2