[8.1.0] Mark built-in providers as hashable (#25230)
Built-in providers always have an immutable `hashCode`. Making them
hashable makes it easier to deduplicate a list of providers, which is
sometimes required as the list of provider instances returned by a rule
must not contain two instances for a given provider.

Starlark-defined providers are already hashable, so this brings the
behavior of built-in providers in line with them.

Closes #24848.

PiperOrigin-RevId: 724370319
Change-Id: I68147b1f707249a8b09ad170c5fc5b3da4776ccf

Commit
https://github.com/bazelbuild/bazel/commit/c94a9bdbaa389f4ed64d895237d443fcca57e8b4

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>