Add ArtifactNestedSetKeys dependencies to rewind graph for propagating actions.
If a propagating action ("mayInsensitivelyPropagateInputs" returns
true) has input artifacts in a nested set, those nested sets are not
invalidated as part of the rewinding if the action needs to be rewound.

This is not currently an issue because the only such action is
SkyframeFilesetManifestAction, which has its own special rewinding logic
with FilesetEntryFunction. However we plan to remove this, so we require
that this case is correctly handled.

Care needs to be taken that we don't break existing code to add
ArtifactNestedSetKeys to the rewind graph, which assumes there are no
such keys in the graph already. The new code to add keys is called near
this process so this interaction is more obvious.

PiperOrigin-RevId: 725983745
Change-Id: I08db9d85f8190a863c21cad17fc2d342c766ef52