[8.2.0] Add `missing_toolchain_error` to the `platform` rule. (#25284)
This allows platforms to provide custom error messages to users when no
required toolchain can be found for a specific type (similar to the
`toolchain_type.no_match_error` from
b60a0f4b4072567e482419fac48c5da91f102648).

This should point to documentation or directions for users to properly
use the platform.

When this is set it replaces the previous message "For more information
on platforms or toolchains see
https://bazel.build/concepts/platforms-intro."

RELNOTES: Add `missing_toolchain_error` to the `platform` rule, to
customize error messages when a required toolchain type cannot be found
for that platform.

Closes #25234.

PiperOrigin-RevId: 725775441
Change-Id: Ica00d215ad4ec8fb0edb714369e1d240d52c08f5