[8.2.0] Preserve insertion order in maps that are traversed (#25288)
Consumers of an `Immutable{List,Map,Set}` cannot know whether it has
been created from a source with unspecified iteration order, such as
`HashMap`. Since `ImmutableMap` itself prescribes that it preserves
insertion order, it's reasonable for consumers to expect a specified,
reasonable, deterministic order. While the current implementation of
`HashMap` in the OpenJDK still results in a deterministic order, this
kind of "sequenced-washing" doesn't cause non-determinism, but it still
runs the risk of having the order of elements change between JDK
versions or vendors.

The locations changed in this PR have been identified by instrumenting
the factory methods of the `Immutable{List,Map,Set}` classes to track
whether they 1) have been created from an unsequenced collection and 2)
are iterated over. I then manually picked out those cases in which I
couldn't rule out that a change in order would be observable to users.

Related to
https://github.com/google/guava/issues/6903#issuecomment-2633372142

Closes #25218.

PiperOrigin-RevId: 724341188
Change-Id: I2f38a4c3dad80055cb8d80853e9f211836019e5b

Commit
https://github.com/bazelbuild/bazel/commit/b03cb636f4091c69c601e6b84f3f4e0a22f1b3ae

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>