Fix extra actions input discovery in CppCompileAction
We only report headers, but not binary precompiled modules.
This results in missing inputs for aspects that need to run over modules with
the `--notrack_incremental_state` flag.

I have tried to share the code with `discoverInputs` as much as possible and keep
the behavior changes to a minimum. However, I have decided to move any state
mutation out of the helper functions I created to make it explicit. This results
in some duplication, but it felt important to keep this visible.

PiperOrigin-RevId: 730984267
Change-Id: I72f47a87fc0038e10437f60b9375959dff356e8f