[8.2.0] Bzlmod: nodep deps (#25381)
This PR implements the "nodep" edges from
https://docs.google.com/document/d/1JsfbH9kdMe3dyOY-IR8SUakS541A7OM8pQcKpxTRMRs/edit?tab=t.0,
using the syntax of `bazel_dep(..., repo_name=None)`. The behavior is
that these edges are "unfulfilled" unless the module they refer to
already exist in the dep graph by some other means.

Most of the changes are in the Discovery class -- I reorganized the code
in there to hopefully help with readability, given the new multi-round
discovery logic.

Also changed `ModuleFileValue.Key` to no longer take the applicable
override next to the module key -- `ModuleFileFunction` now gets the
root module from Skyframe itself and looks up the correct override. The
old setup was always weird (what does it mean to request `foo@1.0` with
an incorrect override??).

Beyond that, I changed `ModuleFileValue` implementations to become
records. Just for the heck of it.

Work towards https://github.com/bazelbuild/bazel/issues/25214

RELNOTES: The `repo_name` parameter of `bazel_dep` can now be set to
`None` to mark it a "nodep" dependency -- that is, the `bazel_dep`
specification is only honored if the target module already exists in the
dependency graph by some other means.

PiperOrigin-RevId: 730962587
Change-Id: I1ca7a7a1228da5e4c2e4b5d6983a2ec97b31a4b8