Fix race condition with multiplex sandboxed workers
Prior to this change, multiplex sandboxed workers shared a working
directory per mnemonic. This caused a race condition when a new
multiplex sandboxed worker with the same mnemonic was launched because
when launching it cleaned the working directory. That could cause
problems for any actions executing in that directory.

This change makes it so each multiplex sandbox worker process has a
unique working directory. It does so while ensuring each
SandboxedWorkerProxy and the associated sandbox are still associated
with the correct multiplexer process and working directory.

Resolves #22589