Make Filesets work when they are created under the frontier.
The reason why they were broken is that the metadata of Filesets were gathered by reaching through the action graph to indirect dependencies, which doesn't work with Skycache, which assumes that direct dependencies of the actions in the active set are enough.

This resulted in the "env.getValue(filesetActionLookupKey)" call in getFilesetOutput() failing with a null pointer exception because ActionLookupKeys are ConfiguredTargetKeys in disguise and the configured target it was referring to was below the frontier and thus not serialized in Skycache.

This is fixed by forwarding said metadata to every artifact that represents a Fileset (its symlink and its symlink tree). This allows one to remove the "reach through the action graph" part of ActionInputMapHelper at the cost of implementing said forwarding.

This also required forwarding this metadata through RunfilesArtifactValue in case there is a Fileset in a runfiles tree.

RELNOTES: None:
PiperOrigin-RevId: 731274463
Change-Id: If4cebd4656fd6506f080998ed8a9edb48f5c5353