Add `use_target_platform_constraints` attribute to the `toolchain` rule.
This causes the toolchain to always match the current target platform, whatever it is. Needed to support test toolchains.

Work towards #25160

Part of #25123.

RELNOTES: Add `use_target_platform_constraints` attribute to the `toolchain` rule, although it is not used, to enable writing custom test toolchains that are valid in 8.2.0 and later releases.

PiperOrigin-RevId: 732123054
Change-Id: I1b65e8881048187fe825bb26812ac6bdfa2b4c16