Build the label to target map without performing skyframe evaluations.
After parsing target patterns we build a `Map<Label, Target>` containing all top-level targets. The code was looking up packages of top-level targets by performing a `MemoizingEvaluator#evaluate` call for each, one at a time, in sequence. A skyframe evaluation is relatively heavyweight given that we can just look up the `PackageIdentifier` in the graph since we already evaluated it on the current build.

PiperOrigin-RevId: 733065574
Change-Id: Ib012c6fd60d8bf8e171b7ab07d4622ed73caaa43