[7.6.0] Actually fix setting the mtime on a shared disk cache. (#25473)
The earlier attempt in
https://github.com/bazelbuild/bazel/commit/e9a18ce24b176f0d9a46af341ed44895c06b7214
was not the correct fix, due to the requirements spelled out in the
utimensat(2) man page: if the current user has write permission but is
not the owner of the file, it may set both atime and mtime to UTIME_NOW,
but any other combination is forbidden. Although in theory we could
still leave the atime alone in other cases, it's not worth the
complexity: atime is often unreliable and Bazel doesn't care about it,
anyway.

Unfortunately, it's difficult to write an integration test for this,
because we can't arrange for the existence of multiple users in our CI
environment.

Fixes https://github.com/bazelbuild/bazel/issues/23512.

PiperOrigin-RevId: 729042375
Change-Id: I7ca5e8bc0858a9796f448fca554d0c9465579e19

Commit
https://github.com/bazelbuild/bazel/commit/e7aef800b313061f982da6ad392de88d5712f9f0

Co-authored-by: Googler <tjgq@google.com>