Globally register BzlLoadValue codec for basic keys
Previously, KeyForBuild and KeyForBuiltins would either be loaded by a DynamicCodec or the LeafObjectCodec, depending on the context. This change now ensures that there is a consistent codec for each subclass of BzlLoadValue.

This is done by splitting BzlLoadValue.Key into two subclasses, KeyForLocalEval and KeyForLocalEval, though this should be opaque to classes depending on BzlLoadValue.Key. KeyForLocalEval's codec, a leaf codec, is now automatically registered. KeyForGlobalEval will still rely on DynamicCodec as it did before.

PiperOrigin-RevId: 733808585
Change-Id: If68cf596ebd0b095c4bb21a9966c6fa9e1d85b13