Introduce packageoids (either packages or package pieces) as containers of targets
Under the old model, targets were owned by packages, target.getPackage()
would get you the package, and target.getPackage().getTargets() would get
you a target's siblings.

Under the new model, targets are owned by packageoids, which can be either
packages or per-macro-instance package pieces.

A packageoid cannot contain a pointer to a package, because that would
defeat the point of lazy macro expansion; we want packages and package
pieces to have different lifetimes, such that invalidating a package piece
does not invalidate sibling package pieces of the same package.

Therefore, in the new model, to go from a packageoid to a package, you
either (1) must be in a special case where it is guaranteed that the
packageoid is a real package (e.g. repo rules), or (2) fetch the package
from skyframe, either directly or through a PackageProvider.

Working towards https://github.com/bazelbuild/bazel/issues/23852

PiperOrigin-RevId: 733849485
Change-Id: Iafb6fc810174775e0d0031a617df0bf7b7eaf3f7