Remove Rule.setContainsErrors() and add a checkState to Package.setContainsErrors()
The only use for Package.setContainsErrors() should be to report an error in
a package that's currently being constructed in situations where the builder
is not available, for example in attribute or rule class machinery. Add a
checkState to Package.setContainsErrors to ensure that is the case, and
call Package.setContainsErrors from Rule.reportError directly.

PiperOrigin-RevId: 734230248
Change-Id: I5aa03df2729ff368a335282e744e3b6dc49e94fc