Move runfiles into ActionExecutionValue.
This dispenses with the need to create RunfilesArtifactValue in ArtifactFunction and running ArtifactFunction for such artifacts.

This requires introducing a new concept, "RichDataProducingAction": if an action that produces anything other than files is an action cache hit, the "anything other" part needs to be re-computed. In theory, this could be serialized in the action cache, but it's currently not so the next best thing is to ask the action to reconstruct it.

Eventually, Filesets will be handled in this way, too, dispensing with the need to special-case them in multiple ways.

RELNOTES: None.
PiperOrigin-RevId: 734804671
Change-Id: I0411c194794f5c267bc7c4017963eca26219f617