Optimize copying outputs out of the sandbox.
When the sandbox and the output tree are in different filesystems (notably, when the latter is in a FUSE filesystem) we fall back to a copy, which is particularly expensive for large directory trees. Instead of relying on FileSystemUtils (which, as explained in comments, does extra work to provide guarantees other callers might need, but sandboxing doesn't), re-implement it as efficiently as possible.

PiperOrigin-RevId: 735359277
Change-Id: I6a3ec3d001e548eb028ad69e0d7938618408f518