Always run Turbine native image with a UTF-8 code page on Windows
We are building Turbine native images with a UTF-8 code page and need to match the runtime code page for UTF-8 paths to be handled properly. This change splits the existing tool for patching the app manifest of the embedded JDK into separate tools for reading and writing app manifests. The write tool is then used to add a manifest to the turbine binary.

Closes #25444.

PiperOrigin-RevId: 735457536
Change-Id: Id6395435cc350772e1225ecb80fa32a74f6ae0e3