[8.8.0] Extend syntax with parameter type annotations
Add allowTypeAnnotations to FileOptions and pass it into Parser. This make it possible to roll out the type annotations gradually / only allowing them in certain locations.
Extend all Parameters with Expression for type.
Parse types on `def` statements.

For the moment types are parsed as general expressions (tests in Python parlance), to keep this change simple. The parsing will be narrowed in a followup change.

Issue: https://github.com/bazelbuild/bazel/issues/24407
PiperOrigin-RevId: 735681498
Change-Id: Iedae56101803d1d29d69c10e4a33ebe7a475ec4c