Refactor lost input ownership calculation into `ActionRewindStrategy` and simplify it.
Lost input ownership calculation is important for action rewinding - in order to rewind properly, we need to know which special artifacts (runfiles, tree artifact, fileset) that a lost input is a part of.

Historically, lost input ownership calculation was performed by `ActionExecutionFunction` by sharing the same code that handles inputs for regular execution. It accomplished this via an interface `ActionInputMapSink` which was implemented by `ActionInputMap` (for regular execution) and `ActionInputDepOwnerMap` (for lost input ownership).

This solution was chosen to guarantee fidelity between the two. However, it is not necessary to accommodate arbitrary ownership mappings between artifacts - there is a well-defined, finite set of possible owner relationships. By using knowledge of the possible owner relationships, we can simplify and improve the performance of the calculation. In this change, we handle the three types of possible owners as follows:

* Runfiles - iterate over all `RunfilesTree`s in the action's inputs and expand them.
* Tree artifact - simply call `hasParent()`/`getParent()` on lost input artifacts.
* Fileset (no change from previous) - assume that fileset ownership is tracked by the spawn strategy that threw the original lost inputs exception. A future change will add fileset ownership handling to `ActionRewindStrategy`, making spawn-strategy-side ownership accounting optional.

Notably, we no longer re-flatten and re-process all of the action's inputs, which should improve rewinding performance.

PiperOrigin-RevId: 735759787
Change-Id: I21fbaea40810cc8c6da79c5d8c6d245318be5252