Fix `TargetSummary` BEP event when a top-level starlark aspect `requires` another aspect.
The `TargetSummaryPublisher` works by anticipating the number of `AspectCompleteEvent` events that will be posted for each target in the build, setting this to a field `aspectCount`. Once the publisher observes `aspectCount + 1` build statuses from `TargetCompleteEvent`/`AspectCompleteEvent`s for a given target, it computes the `TargetSummary`.

This logic was broken in the case where a top-level starlark aspect `requires` another aspect. That logic was broken because the `requires`-ed aspect is considered a top-level aspect but the `TargetSummaryPublisher` was only expecting the aspects listed in `--aspects`.

In the new approach, the `TargetSummaryPublisher` listens for a new event containing the number of all top-level aspects, including those recursively `requires`-ed.

PiperOrigin-RevId: 736273988
Change-Id: I2bd0d703f0a976f8d3763d764b43244c145ab2a1