[8.2.0] Refactor build rewinding to a fallback strategy for action rewinding (#25545)
By throwing `LostInputExecExceptions` in the right locations, build
rewinding can be realized as a coarse-grained fallback for the more
fine-grained action rewinding. This approach avoids further divergence
between Bazel and Blaze logic and ensures that further improvements to
build rewinding also simplify future efforts to add action rewinding to
Bazel.

This should be a pure refactoring with one exception: if the download of
the outputs of a remotely executed action (with caching disabled) fails
due cache eviction, the build is no longer retried after the remote
execution has already been retried. An RE backend that repeatedly fails
to keep the outputs of a freshly excuted action in the cache long enough
for them to be downloaded is faulty and additional full build retries
are pointless after retries of the action execution have failed
repeatedly. This change in behavior is required to get existing tests to
pass.

Closes #25396.

PiperOrigin-RevId: 734550819
Change-Id: Idabee2483beb05721c413c27ab610b8f260412ec 
(cherry picked from commit e8526d4ff3bcdf30507a551e824623e38b50e7c5)

Closes #25544