Remove "inputs of interest" from `ActionInputDepOwnerMap`.
This was used to filter the stored ownership mappings down to just artifacts that were actually lost, but the new ownership calculation strategy can do this itself.

Many places now create an empty `ActionInputDepOwnerMap` and never add anything to it. I will leave cleaning up those for a followup, once it's optional to pass one.

Also add a method `RunfilesArtifactValue#getAllArtifacts` since that's all `ActionRewindStrategy` needs (not the metadata).

PiperOrigin-RevId: 736656703
Change-Id: I898afe3495a5c5a54c3e3a682b8ca72703a3d240