Add and set `--guard_against_concurrent_changes=lite`
Without `--guard_against_concurrent_changes`, locally executed actions can easily pollute the disk or remote cache when inputs are modified during execution, which is in stark contrast to Bazel's correctness guarantees.
But with the flag enabled, the additional file system operations can unnecessarily slow down non-interactive builds and also, in rare but not negligible cases, result in false positives, which has made it difficult to flip the flag.

This change aims to resolve this issue by introducing a "lite" value of the flag that only checks source files in the main repository. Since output files are made read-only after digesting and files in external repositories are rarely modified during execution (`local_repository` being a notable exception), this gets most of the benefits for correctness while requiring far fewer stats (99% fewer calls to `wasModifiedSinceDigest` on the Bazel project). It also avoids the false positives that can arise for output files.

Along the way, this change unifies concurrent change detection between `RemoteSpawnCache` and `RemoteSpawnRunner` by moving it into `RemoteExecutionService#uploadOutputs`. The implementation in `RemoteSpawnRunner` required to stats per input and the `RemoteSpawnCache` implementation unnecessarily retained the spawn input mapping, both of which is fixed in the new, centralized implementation.

# Conflicts:
#	src/main/java/com/google/devtools/build/lib/remote/RemoteSpawnCache.java