Fix manifest Skycache violation checks
Fixes issues like violations being issued when the baseline file was identical to the client file. This can happen because the manifest doesn't always get updated, as it's the case after reverting a file with Fig. This change also adds code to issue instructions to clean up the workspace whenever we encounter violations.

RELNOTES:none
PiperOrigin-RevId: 737543981
Change-Id: I1353ce7e9efb396c38bcb9221d3cafc181bd6266