Make lost input owners tracking optional for spawn runners.
If not provided, `ActionRewindStrategy` can calculate the owners. Thus, the only reason to provide them is for efficiency.

Note that owners are still required for lost outputs, at least for now. It's not as easy to make them optional, because `CompletionFunction` needs to remove lost output owners from the set of successfully built artifacts.

PiperOrigin-RevId: 737622448
Change-Id: I9ce2ab518afa0ba2e54efb2ac34248d5996965c8