Rebrand `ActionInputDepOwners` as `LostInputOwners`.
This is used strictly for rewinding, so name it more specifically and move it to the `rewinding` package.

Additionally, instead of an interface, make it a concrete class so that every tracker of lost input owners is forced to go through the validity checks added in https://github.com/bazelbuild/bazel/commit/76145b8564f0167beda1ed0f91aca043ea34c8ac.

PiperOrigin-RevId: 737759745
Change-Id: Ie986bf3483b74250b91b0dfdee1f13c27ddeca12