Enable some C++ `RewindingTest`s in Bazel
This relies on a test toolchain based on the default Unix toolchain. Since `RewindingTest` disables transitive package tracking, a symlink to `bazel_tools` under the execroot has to be created manually to mimic disabled Skymeld behavior.

While an almost identical macOS toolchain could be added, it wouldn't work with the rewinding test cases due to the lack of support for nodeps dynamic libraries and `--start-lib/--end-lib`.

Closes #25501.

PiperOrigin-RevId: 737890619
Change-Id: I1fd597285030e211645092f8acc2c2565acbaa72