Do not add individual TreeFileArtifacts to the set of allowed inputs.
Things seem to work fine without this, and the metadata of the tree artifact is not available if the tree artifact is a declared input (i.e. a scheduling dependency), is not returned by include scanning but a file in it is used by the compiler.

The goal is to eventually remove ArtifactExpander. Work on other call sites is ongoing.

RELNOTES: None.
PiperOrigin-RevId: 738003841
Change-Id: I17d191bf2e85a573e610edbda11a973be77dcb2b