[8.2.0] Prevent concurrent use of the same `CharsetEncoder` in `json.decode` (#25612)
Encoders are stateful, and `json.decode` may be called concurrently by
multiple threads. This will (if we're lucky) manifest as an
`IllegalStateException` thrown by `CharsetEncoder.encode()`.

We could consider keeping a single instance around for reuse (i.e. a
one-element object pool), but I'll refrain from doing so until we have
evidence that it's worth optimizing.

PiperOrigin-RevId: 737594384
Change-Id: I8d8c87a9919661ef49490293f54d513ee1fe6b69