Cleanup handling of SpawnAction.executionInfo in serialization.
It's incorrect for AutoCodec to use getExecutionInfo to retrieve this
because that may change the content. It's also a bit type unsafe
that the field is marked ImmutableMap but needs to deserialize
as an ImmutableSortedMap.  Fixes these issues by renaming the
underyling field sortedExecutionInfo, marking its type precisely as
ImmutableSortedMap<String, String> and referencing it from codecs.

PiperOrigin-RevId: 738380105
Change-Id: I3a6bb4246a6f941b4b615fb7dca5308667886a67