Make Skycache work again in the wake of https://github.com/bazelbuild/bazel/commit/25dba3a109b6c3c0038b8da03e81c7538ce2e121.
That change marked RunfilesTreeAction as not shareable, which, as an unexpected side effect, made every such action part of the active set and since the upwards transitive closure of an active node is also active, made every node active that depends on a runfiles tree, which is almost all of them since almost every node depends on the execution of at least one tool.

This is fixed by changing the algorithm to determine the active set to directly check whether a node depends on the build ID.

RELNOTES: None.
PiperOrigin-RevId: 738392181
Change-Id: I15a142a9d16bb07ab3d5d997aa7915e865db92b7