Pass the multiplexer log file to sandboxed worker proxy instead of using a new log file.
Prior to this change, we create different log files for different sandbox proxies. That results in a FileNotFound exception because a multiplex worker is a single process which pipes the stderr stream to the logfile, and different proxies in the same process is trying to access different files.

With this change, all proxies point to the same file, which is referenced when the instance of the multiplexer is created.

PiperOrigin-RevId: 738476002
Change-Id: I397b0de9efdb6821dea9e5b37c605a9924d55d34