Disambiguate anchor links on the CLI reference page
As noted in
[bazelbuild/bazel#14822](https://github.com/bazelbuild/bazel/issues/14822),
on the [command-line reference
page](https://bazel.build/reference/command-line-reference), each
option flag is assigned an ID tag identical to the flag name. Because
some flags appear under more than one command, the duplicated IDs
sometimes make it impossible to link directly to a specific command’s
flag.

This PR prefixes each flag’s ID tag with the command name, ensuring
uniqueness across the page. To maintain backwards compatibility, we
move the old (ambiguous) IDs to the <code> element of the
documentation, so that existing links still work as before.

This change was tested by running `bazel build
//src/main/java/com/google/devtools/build/lib:gen_command-line-reference`. Here
is the relevant output for the `--[no]build` flag.

```html
<dt id="build-flag--build"><code id="build"><a href="#build-flag--build">--[no]build</a></code> default: "true"</dt>
```