Delete RPCServer interface
See unknown commit for why this existed. It survived that change because it could
be useful to keep GrpcServerImpl from getting unwiedly, but also because I
didn't feel like cleaning it up at that time. The former is maybe still true,
the latter less so, since it created some extra busy work while I was
prototyping something.

PiperOrigin-RevId: 739169753
Change-Id: Ic7ac87f59dd8c7266737018e542598d7c3d6834b