[8.2.0] Fix race condition with multiplex sandboxed workers (#25423)
Prior to this change, multiplex sandboxed workers shared a working
directory per mnemonic. This caused a race condition when a new
multiplex sandboxed worker with the same mnemonic was launched because
when launching it cleaned the working directory. That could cause
problems for any actions executing in that directory.

This change makes it so each multiplex sandbox worker process has a
unique working directory. It does so while ensuring each
SandboxedWorkerProxy and the associated sandbox are still associated
with the correct multiplexer process and working directory.

Resolves #22589

I couldn't figure out if the Bazel repo has an autoformatter somewhere,
so I did my best to manually format the code with a style that follows
the existing code.

Closes #25400.

PiperOrigin-RevId: 732256101
Change-Id: If8deea240fda77780feaeac352cf099fb9bfcee3

Commit
https://github.com/bazelbuild/bazel/commit/d54fc621563ca501cee11dbff1c46649ff4b7e55

Co-authored-by: jjudd <james@lab-y.com>