[8.2.0] Allow finalizer targets to see targets visible either to BUILD file or to finalizer macro class definition
Previously, finalizers used the same visibility semantics as normal symbolic
macros. As a result, migrating an epilogue legacy macro to a finalizer could
result in visibility errors if, for example, a finalizer-defined target
declared a dependency on a legacy-macro-defined target which was visible to
the BUILD file but not to the finalizer's definition.

PiperOrigin-RevId: 739300166
Change-Id: I480dde3d2a9704ef85a64ba8346038e784172ee6

Commit
https://github.com/bazelbuild/bazel/commit/4c2d91e762ab6e492853b021408129dd93fb5904