Include `DefaultInfo` in `cquery` `providers` dict
All targets observed from Starlark advertise the `DefaultInfo` provider, but internally that provider is split into `FilesProvider`, `FilesToRunProvider` and `RunfilesProvider`. This is confusing to users that use `cquery` to inspect targets using Starlark expressions. This is improved by synthesizing a `DefaultInfo` provider in the same way as during regular analysis. As a result, the `default_runfiles` and `data_runfiles` of a target can now be inspected with `cquery`.