Tolerate journal corruption in the action cache whenever it doesn't compromise referential integrity.
This removes the need for the PersistentMap.load() parameter and clarifies the handling of exceptional conditions during loading.

The journal file is updated in a fundamentally unsafe manner: abnormal termination can fail to flush the file and leave incomplete contents behind. Truncation that merely loses the most recent updates is safe to ignore; truncation that results in loss of referential integrity is still detected through the validation record.

In addition, when corruption is detected, move the corrupted files into a separate directory. This way, the last seen corrupted state stays around even if the action cache directory is deleted and recreated during garbage collection (to be implemented in a followup).

PiperOrigin-RevId: 740783332
Change-Id: I784ab4088e469bf15c9731d6f45b6843e6203269