Make the optimization to skip serializing generating action keys less brittle.
When an `ArtifactSerializationContext` indicates that serialization of a generating action key should be omitted, it is responsible for providing the omitted generating action key during deserialization. This way, we only create fully functional deserialized `Artifact` instances instead of partially functional instances containing the `OMITTED_FOR_SERIALIZATION` marker object (which is now deleted).

PiperOrigin-RevId: 740849784
Change-Id: Iadbc1e01a3e81a92169f0e23f78d4acd6c39a095