All `FilesetOutputSymlink`s point to an `Artifact`.
Recent restrictions added to the fileset rule API mean that all entries point to an `Artifact` that is already in the build. Thus, we can simplify code and save memory by storing the `Artifact` instead of an arbitrary path string.

There are now many scenarios in which an `ActionInput` reference is actually guaranteed to be an `Artifact`, making followup simplifications in input metadata handling and action rewinding possible.

PiperOrigin-RevId: 740858578
Change-Id: I21c198a1e19d3ec570d387c2339da6a1c0bc79b8